approach-1-keep-alive-manually/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
./azure_eventhub-5.10.0b1-py3-none-any.whl